Add sub-task persistence
