+++ app/routers/tasks.py
@@ -1,91 +1,120 @@
-from fastapi import APIRouter, HTTPException, status
 from typing import List, Optional
 from uuid import UUID
-from app.models import Task  # Assuming models.py is in the app directory
+from fastapi import APIRouter, Depends, HTTPException, status
+from sqlalchemy.orm import Session
+from ..models import Task
+from ..db import SessionLocal, Base, engine, TaskORM
 
-router = APIRouter(
-    prefix="/tasks",
-    tags=["tasks"],
-)
+# Ensure tables exist
+Base.metadata.create_all(bind=engine)
 
-tasks_db: List[Task] = []
-
-
-# Helper function to find a task recursively
-def _find_task_recursive(task_id: UUID,
-                         task_list: List[Task]) -> Optional[Task]:
-    for task in task_list:
-        if task.id == task_id:
-            return task
-        if task.sub_tasks:
-            found_in_sub = _find_task_recursive(task_id, task.sub_tasks)
-            if found_in_sub:
-                return found_in_sub
-    return None
+router = APIRouter(prefix="/tasks", tags=["tasks"])
 
 
-# Helper function to delete a task recursively
-def _delete_task_recursive(task_id: UUID, task_list: List[Task]) -> bool:
-    for i, task in enumerate(task_list):
-        if task.id == task_id:
-            task_list.pop(i)
-            return True
-        if task.sub_tasks:
-            if _delete_task_recursive(task_id, task.sub_tasks):
-                return True
-    return False
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+
+def _create_task_recursive(task: Task, db: Session, parent_id: Optional[str] = None) -> Task:
+    task_orm = TaskORM(
+        title=task.title,
+        description=task.description,
+        due_date=task.due_date,
+        priority=task.priority,
+        completed=task.completed,
+        parent_id=parent_id,
+    )
+    db.add(task_orm)
+    db.commit()
+    db.refresh(task_orm)
+    sub_tasks = [
+        _create_task_recursive(st, db, parent_id=task_orm.id)
+        for st in task.sub_tasks
+    ]
+    return Task(
+        id=UUID(task_orm.id),
+        title=task_orm.title,
+        description=task_orm.description,
+        due_date=task_orm.due_date,
+        priority=task_orm.priority,
+        completed=task_orm.completed,
+        sub_tasks=sub_tasks,
+    )
+
+
+def _build_task_recursive(task_orm: TaskORM, db: Session) -> Task:
+    children = (
+        db.query(TaskORM).filter(TaskORM.parent_id == task_orm.id).all()
+    )
+    sub_tasks = [_build_task_recursive(ch, db) for ch in children]
+    return Task(
+        id=UUID(task_orm.id),
+        title=task_orm.title,
+        description=task_orm.description,
+        due_date=task_orm.due_date,
+        priority=task_orm.priority,
+        completed=task_orm.completed,
+        sub_tasks=sub_tasks,
+    )
+
+
+def _delete_sub_tasks(task_orm: TaskORM, db: Session) -> None:
+    children = db.query(TaskORM).filter(TaskORM.parent_id == task_orm.id).all()
+    for ch in children:
+        _delete_sub_tasks(ch, db)
+        db.delete(ch)
 
 
 @router.post("/", response_model=Task, status_code=status.HTTP_201_CREATED)
-async def create_task(task: Task):
-    # The Task model already assigns a UUID if not provided.
-    # It also handles default for sub_tasks.
-    # We need to ensure any provided sub_tasks also get IDs if they don't
-    # have them. This is complex if sub_tasks can have their own sub_tasks
-    # deeply nested. For now, Pydantic handles the creation of sub_task models.
-    tasks_db.append(task)
-    return task
+def create_task(task: Task, db: Session = Depends(get_db)):
+    return _create_task_recursive(task, db)
 
 
 @router.get("/", response_model=List[Task])
-async def read_tasks():
-    return tasks_db
+def read_tasks(db: Session = Depends(get_db)):
+    tasks = db.query(TaskORM).filter(TaskORM.parent_id == None).all()
+    return [_build_task_recursive(t, db) for t in tasks]
 
 
 @router.get("/{task_id}", response_model=Task)
-async def read_task(task_id: UUID):
-    task = _find_task_recursive(task_id, tasks_db)
-    if task is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail="Task not found")
-    return task
+def read_task(task_id: UUID, db: Session = Depends(get_db)):
+    task = db.query(TaskORM).filter(TaskORM.id == str(task_id)).first()
+    if not task:
+        raise HTTPException(status_code=404, detail="Task not found")
+    return _build_task_recursive(task, db)
 
 
 @router.put("/{task_id}", response_model=Task)
-async def update_task(task_id: UUID, task_update: Task):
-    task = _find_task_recursive(task_id, tasks_db)
-    if task is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail="Task not found")
+def update_task(task_id: UUID, task_update: Task, db: Session = Depends(get_db)):
+    task = db.query(TaskORM).filter(TaskORM.id == str(task_id)).first()
+    if not task:
+        raise HTTPException(status_code=404, detail="Task not found")
 
-    # Update fields, ensuring ID from path is respected
     task.title = task_update.title
     task.description = task_update.description
     task.due_date = task_update.due_date
     task.priority = task_update.priority
     task.completed = task_update.completed
-    task.sub_tasks = task_update.sub_tasks  # Direct replacement for now
+    _delete_sub_tasks(task, db)
+    db.commit()
+    for st in task_update.sub_tasks:
+        _create_task_recursive(st, db, parent_id=task.id)
+    db.commit()
+    db.refresh(task)
 
-    # The ID of the task itself should not change to task_update.id
-    # task.id remains task_id from the path
-
-    return task
+    return _build_task_recursive(task, db)
 
 
 @router.delete("/{task_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_task(task_id: UUID):
-    deleted = _delete_task_recursive(task_id, tasks_db)
-    if not deleted:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail="Task not found")
-    return  # No content
+def delete_task(task_id: UUID, db: Session = Depends(get_db)):
+    task = db.query(TaskORM).filter(TaskORM.id == str(task_id)).first()
+    if not task:
+        raise HTTPException(status_code=404, detail="Task not found")
+    _delete_sub_tasks(task, db)
+    db.delete(task)
+    db.commit()
+    return
