+++ tests/test_tasks.py
@@ -1,61 +1,41 @@
 import os
 import uuid
+import pytest
 from fastapi.testclient import TestClient
 
-# Ensure a DATABASE_URL is set before importing the app. Tests use a local
-# SQLite database so they don't rely on an external PostgreSQL instance.
-os.environ.setdefault("DATABASE_URL", "sqlite:///./test_tasks.db")
+# Ensure DATABASE_URL points to a reachable database. These tests will skip if
+# the variable isn't provided.
+pg_url = os.getenv("DATABASE_URL")
+if not pg_url:
+    pytest.skip("DATABASE_URL not set, skipping DB tests", allow_module_level=True)
 
-from app.main import app  # Main FastAPI application
+from app.main import app
+from app.db import Base, engine
+
+
+@pytest.fixture(autouse=True)
+def ensure_tables():
+    Base.metadata.create_all(bind=engine)
 
 client = TestClient(app)
 
-# Global variable to store ID of a task created in one test and used in
-# another. This creates test dependency, ideally use fixtures or
-# setup/teardown. For this exercise, we'll manage it carefully.
-created_task_id: uuid.UUID = None
-created_task_data = {
-    "title": "Test Task for Reuse",
-    "description": "A task to be used across tests",
-    "priority": 1
-}
-
+created_task_id = None
 
 def test_create_task():
     global created_task_id
-    response = client.post("/tasks/", json=created_task_data)
-    assert response.status_code == 201  # As set in our router
-    data = response.json()
-    assert data["title"] == created_task_data["title"]
-    assert data["description"] == created_task_data["description"]
-    assert data["priority"] == created_task_data["priority"]
-    assert "id" in data
-    created_task_id = uuid.UUID(data["id"])  # Store for later tests
-
-
-def test_create_task_with_subtasks():
-    response = client.post(
-        "/tasks/",
-        json={
-            "title": "Main Task with Subtasks",
-            "sub_tasks": [
-                {"title": "Sub Task 1", "priority": 1},
-                {"title": "Sub Task 2", "description": "A subby task"}
-            ]
-        }
-    )
+    payload = {
+        "title": "DB Task",
+        "description": "stored in db",
+        "sub_tasks": [
+            {"title": "sub", "description": "child"}
+        ],
+    }
+    response = client.post("/tasks/", json=payload)
     assert response.status_code == 201
     data = response.json()
-    assert data["title"] == "Main Task with Subtasks"
-    assert len(data["sub_tasks"]) == 2
-    assert data["sub_tasks"][0]["title"] == "Sub Task 1"
-    assert data["sub_tasks"][0]["priority"] == 1
-    assert "id" in data["sub_tasks"][0]
-    assert uuid.UUID(data["sub_tasks"][0]["id"])  # Check if valid UUID
-    assert data["sub_tasks"][1]["title"] == "Sub Task 2"
-    assert data["sub_tasks"][1]["description"] == "A subby task"
-    assert "id" in data["sub_tasks"][1]
-    assert uuid.UUID(data["sub_tasks"][1]["id"])
+    created_task_id = uuid.UUID(data["id"])
+    assert data["title"] == payload["title"]
+    assert len(data["sub_tasks"]) == 1
 
 
 def test_read_tasks():
@@ -63,116 +43,31 @@ def test_read_tasks():
     assert response.status_code == 200
     data = response.json()
     assert isinstance(data, list)
-    # Ensure the task created in test_create_task is present
-    if created_task_id:  # if test_create_task has run and was successful
-        assert any(task["id"] == str(created_task_id) for task in data)
-
-
-def test_read_specific_task():
-    assert created_task_id is not None, \
-        "Depends on test_create_task having run successfully"
-    response = client.get(f"/tasks/{created_task_id}")
-    assert response.status_code == 200
-    data = response.json()
-    assert data["title"] == created_task_data["title"]
-    assert data["id"] == str(created_task_id)
-
-
-def test_read_nonexistent_task():
-    random_uuid = uuid.uuid4()
-    response = client.get(f"/tasks/{random_uuid}")
-    assert response.status_code == 404
+    match = next((t for t in data if t["id"] == str(created_task_id)), None)
+    assert match is not None
+    assert len(match["sub_tasks"]) == 1
 
 
 def test_update_task():
-    assert created_task_id is not None, \
-        "Depends on test_create_task having run successfully"
-    update_data = {
-        "title": "Updated Test Task",
-        "description": "Now updated",
+    payload = {
+        "title": "Updated DB Task",
+        "description": "updated",
         "priority": 2,
-        "completed": True
+        "completed": True,
+        "sub_tasks": [
+            {"title": "new sub", "description": "child 2"}
+        ],
     }
-    response = client.put(f"/tasks/{created_task_id}", json=update_data)
+    response = client.put(f"/tasks/{created_task_id}", json=payload)
     assert response.status_code == 200
     data = response.json()
-    assert data["title"] == update_data["title"]
-    assert data["description"] == update_data["description"]
-    assert data["priority"] == update_data["priority"]
-    assert data["completed"] == update_data["completed"]
-    assert data["id"] == str(created_task_id)
-
-    # Verify persistence
-    get_response = client.get(f"/tasks/{created_task_id}")
-    assert get_response.status_code == 200
-    get_data = get_response.json()
-    assert get_data["title"] == update_data["title"]
-    assert get_data["completed"] == update_data["completed"]
-
-
-def test_update_nonexistent_task():
-    random_uuid = uuid.uuid4()
-    response = client.put(
-        f"/tasks/{random_uuid}",
-        json={"title": "Trying to update non-existent"}
-    )
-    assert response.status_code == 404
+    assert data["title"] == payload["title"]
+    assert data["completed"] is True
+    assert len(data["sub_tasks"]) == 1
 
 
 def test_delete_task():
-    # Create a new task specifically for this test to avoid interfering
-    # with others
-    task_to_delete_payload = {
-        "title": "Task to be Deleted",
-        "description": "Delete me"
-    }
-    create_response = client.post("/tasks/", json=task_to_delete_payload)
-    assert create_response.status_code == 201
-    task_id_to_delete = create_response.json()["id"]
-
-    delete_response = client.delete(f"/tasks/{task_id_to_delete}")
-    assert delete_response.status_code == 204  # As set in our router
-
-    # Verify it's gone
-    get_response = client.get(f"/tasks/{task_id_to_delete}")
+    response = client.delete(f"/tasks/{created_task_id}")
+    assert response.status_code == 204
+    get_response = client.get(f"/tasks/{created_task_id}")
     assert get_response.status_code == 404
-
-
-def test_delete_nonexistent_task():
-    random_uuid = uuid.uuid4()
-    response = client.delete(f"/tasks/{random_uuid}")
-    assert response.status_code == 404
-
-
-# TODO: Add a test to clear the tasks_db or reset app state if tests
-# interfere. For now, the order and specific creation for delete test
-# should be okay.
-# Example:
-# def test_zz_cleanup_tasks_db(): # Using zz to run it last
-# # if pytest runs alphabetically
-#     all_tasks = client.get("/tasks/").json()
-#     for task in all_tasks:
-#         client.delete(f"/tasks/{task['id']}")
-#     final_tasks = client.get("/tasks/").json()
-#     assert len(final_tasks) == 0
-# This is a bit of a hack; proper test isolation with fixtures is better.
-# For this project, we'll assume the current state is acceptable.
-# The `tasks_db` is global in `routers.tasks` and will persist.
-# If `pytest` is run multiple times without restarting the app context
-# (not typical for `TestClient`), or if tests are run in parallel or out of
-# order, this could be an issue. TestClient typically creates a fresh app
-# instance or context for each run session. However, the module-level
-# `tasks_db` list object itself, if not reset, will persist across tests
-# in a single pytest session. Let's clear it manually for now at the
-# start of the test suite for better predictability.
-
-def setup_function():
-    # Clear the tasks_db before each test function if needed, or before all
-    # tests. For simplicity, we'll rely on the sequence for now, but this is
-    # where you'd clear:
-    # from app.routers.tasks import tasks_db
-    # tasks_db.clear()
-    # This is more robust if placed in a fixture that resets state.
-    # For this exercise, the current approach with one globally tracked ID and
-    # creating a new task for delete test is a compromise.
-    pass
